# shared/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: HabitCraft API
  description: |
    REST API specification for the HabitCraft application.
    All backend implementations should follow this specification.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.habitcraft.example.com/v1
    description: Production server
...
paths:
...
  /habits/{habitId}/completions:
    get:
      tags:
        - Completions
      summary: List completions
      description: Get all completions for a habit
      operationId: listCompletions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/HabitId'
        - name: startDate
          in: query
          description: Filter completions from this date
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter completions until this date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of completions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Completion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Completions
      summary: Mark habit complete
      description: Mark a habit as completed for a specific date
      operationId: createCompletion
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/HabitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - date
              properties:
                date:
                  type: string
                  format: date
                  example: "2024-01-15"
                notes:
                  type: string
                  example: "Went for a 30-minute run"
      responses:
        '201':
          description: Completion recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Completion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Already completed for this date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /habits/{habitId}/completions/{date}:
    put:
      tags:
        - Completions
      summary: Update completion note
      description: Update the note for an existing habit completion
      operationId: updateCompletionNote
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/HabitId'
        - name: date
          in: path
          required: true
          description: Date of the completion to update
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                  nullable: true
                  maxLength: 500
                  example: "Went for a 45-minute run in the park"
                  description: Note content (pass null to clear the note)
      responses:
        '200':
          description: Completion note updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Completion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Completions
      summary: Remove completion
      description: Remove a habit completion for a specific date
      operationId: deleteCompletion
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/HabitId'
        - name: date
          in: path
          required: true
          description: Date of the completion to remove
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Completion removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  habitId:
                    type: string
                  date:
                    type: string
                    format: date
              example:
                message: "Completion deleted successfully"
                habitId: "abc123"
                date: "2024-01-15"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    HabitId:
      name: habitId
      in: path
      required: true
      description: Habit identifier
      schema:
        type: string
        format: uuid
  schemas:
...
    Completion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        habitId:
          type: string
          format: uuid
        date:
          type: string
          format: date
          example: "2024-01-15"
        notes:
          type: string
          example: "Went for a 30-minute run"
        createdAt:
          type: string
          format: date-time
      example:
        id: "abc123"
        habitId: "habit-456"
        date: "2024-01-15"
        notes: "Went for a 30-minute run"
        createdAt: "2024-01-15T08:30:00Z"
    Error:
      type: object
      properties:
        error:
          type: string
          example: Resource not found
        message:
          type: string
          example: The requested habit does not exist
        statusCode:
          type: integer
          example: 404
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - access denied to this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Access denied"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'